backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
langgraph==0.2.45
openai==1.54.0
